feat(auth): Implement user registration and role management
### Changes:
- **Added `/auth/signup` endpoint**:
  - Registers new users with a username, email, and password.
  - Hashes passwords before saving.
  - Assigns roles based on the request (default is `ROLE_USER`).
  - Checks for duplicate usernames and emails before registration.
  - Returns a message response upon successful registration.

- **Updated `/auth/signin` endpoint**:
  - Authenticates users using `AuthenticationManager`.
  - Generates JWT token upon successful login.
  - Returns user details, roles, and JWT in the response.

- **Introduced `RoleRepository`**:
  - Provides methods to fetch roles from the database.

- **Created `MessageResponse` class**:
  - Standardizes API responses with a message field.

- **Added `SignupRequest` DTO**:
  - Handles user registration requests with validation.

- **Introduced `Role` entity**:
  - Defines role attributes and mappings in the database.

### Impact:
- Enables secure user registration and authentication.
- Supports role-based access control with flexible role assignment.
- Provides structured API responses for authentication requests.

+++ src/main/java/uk/gegc/ecommerce/sbecom/model/Role.java
@@ -14,7 +14,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    private int roleID;
+    private Long roleID;
 
     @ToString.Exclude
     @Enumerated(EnumType.STRING)
